requirements.txt
langchain
streamlit
python-dotenv
pdf2image
pillow
google-generativeai